refactor/improve VTEC jabber message generation
closes #614
