+++ data/product_examples/WCN/WCNLIX.txt
@@ -1,0 +1,84 @@
+128 
+WWUS64 KLIX 102246
+WCNLIX
+
+WATCH COUNTY NOTIFICATION FOR WATCHES 343/344
+NATIONAL WEATHER SERVICE NEW ORLEANS LA
+546 PM CDT FRI JUN 10 2022
+
+LAC005-063-091-093-095-105-110000-
+/O.CAN.KLIX.SV.A.0344.000000T0000Z-220611T0200Z/
+
+THE NATIONAL WEATHER SERVICE HAS CANCELLED SEVERE THUNDERSTORM
+WATCH 344 FOR THE FOLLOWING AREAS
+
+IN LOUISIANA THIS CANCELS 6 PARISHES
+
+IN SOUTHEAST LOUISIANA
+
+ASCENSION             LIVINGSTON            ST. HELENA           
+ST. JAMES             ST. JOHN THE BAPTIST  TANGIPAHOA           
+
+THIS INCLUDES THE CITIES OF AMITE, CONVENT, DARLINGTON, 
+DENHAM SPRINGS, DONALDSONVILLE, EASLEYVILLE, GONZALES, GRAMERCY, 
+GREENSBURG, HAMMOND, LAPLACE, LUTCHER, MONTPELIER, PONCHATOULA, 
+PRAIRIEVILLE, RESERVE, ROBERT, WALKER, AND WATSON.
+
+$$
+
+GMZ530-110000-
+/O.CAN.KLIX.SV.A.0344.000000T0000Z-220611T0200Z/
+
+THE NATIONAL WEATHER SERVICE HAS CANCELLED SEVERE THUNDERSTORM
+WATCH 344 FOR THE FOLLOWING AREAS
+
+THIS CANCELS THE FOLLOWING ADJACENT COASTAL WATERS
+
+LAKE PONTCHARTRAIN AND LAKE MAUREPAS        
+
+$$
+
+LAC007-057-109-110200-
+/O.CON.KLIX.SV.A.0344.000000T0000Z-220611T0200Z/
+
+SEVERE THUNDERSTORM WATCH 344 REMAINS VALID UNTIL 9 PM CDT THIS
+EVENING FOR THE FOLLOWING AREAS
+
+IN LOUISIANA THIS WATCH INCLUDES 3 PARISHES
+
+IN SOUTHEAST LOUISIANA
+
+ASSUMPTION            LAFOURCHE             TERREBONNE           
+
+THIS INCLUDES THE CITIES OF BAYOU CANE, CUT OFF, GALLIANO, HOUMA,
+LABADIEVILLE, LAROSE, PAINCOURTVILLE, PIERRE PART, RACELAND, 
+AND THIBODAUX.
+
+$$
+
+GMZ550-110200-
+/O.CON.KLIX.SV.A.0344.000000T0000Z-220611T0200Z/
+
+SEVERE THUNDERSTORM WATCH 344 REMAINS VALID UNTIL 9 PM CDT THIS
+EVENING FOR THE FOLLOWING AREAS
+
+THIS WATCH INCLUDES THE FOLLOWING ADJACENT COASTAL WATERS
+
+COASTAL WATERS FROM PORT FOURCHON LA TO LOWER ATCHAFALAYA RIVER
+LA OUT 20 NM            
+
+$$
+
+GMZ536-557-110100-
+/O.CON.KLIX.SV.A.0343.000000T0000Z-220611T0100Z/
+
+SEVERE THUNDERSTORM WATCH 343 REMAINS VALID UNTIL 8 PM CDT THIS
+EVENING FOR THE FOLLOWING AREAS
+
+THIS WATCH INCLUDES THE FOLLOWING ADJACENT COASTAL WATERS
+
+CHANDELEUR SOUND      
+COASTAL WATERS FROM PASCAGOULA MISSISSIPPI TO STAKE ISLAND OUT
+20 NM                    
+
+$$
